big commit: API CHANGES step_by has been 'drawn in' to step.  now step() is equivilent to step(1000/60) by default
TEST CHANGES

Took out the dodgy test that expected the aggregate of the timedeltas to be the same as the jasmine waits(1000), within 10%.  It's subject to breaking when functioning correctly due to a particularly fast engine calling requestAnimationFrame too many times.
Replaced the *uber* dodgy method of blocking with an actual sleep function.

INTERNAL CHANGES
Quite a lot of refactoring, especially within _spin_flywheel.  This will need to be tested against previous iterations, clearly